feature_serving/pyproject.toml
[tool.poetry]
name = "feature-service" # Or whatever name you chose
version = "0.1.0"
description = "AgriYield Feature Service to provide weather data from Swift"
authors = ["Your Name <your.email@example.com>"] # Or empty
readme = "README.md" # Optional
package-mode = false

[tool.poetry.dependencies]
python = "^3.10" # Or the version you chose
fastapi = "^0.104.1" # Or the version you installed
uvicorn = {extras = ["standard"], version = "^0.23.2"} # Use standard extra for dependencies like httptools
pandas = "^2.1.2" # Or the version you installed
pydantic-settings = "^2.0.3" # Or the version you installed
structlog = "^23.2.0" # Or the version you installed
openstacksdk = "^1.2.0" # Or the version you installed


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"